requirements.txt
fastapi==0.111.1
SQLAlchemy==2.0.31
uvicorn==0.30.1
alembic==1.13.2
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.9
python-jose[cryptography]
PyJWT==2.8.0